requirements.txt
pyshark
fastapi
uvicorn
python-multipart
python-nmap
nest-asyncio
requests